elite_backend/requirements.txt
Django==5.0.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-rest-passwordreset
Pillow==10.1.0
python-decouple==3.8 